build: move desktop app to its own directory

+++ desktop/dist.sh
@@ -1,13 +1,22 @@
 #!/bin/bash
+
+EXTENSION_PATH="${PWD}/../extension/.output/chrome-mv3"
+
+if [ ! -f "$EXTENSION_PATH/manifest.json" ]; then
+  echo "ERROR: extension not found in $EXTENSION_PATH"
+  exit 1
+fi
+
 docker run --rm -ti \
   --env-file <(env | grep -iE 'DEBUG|NODE_|ELECTRON_|YARN_|NPM_|CI|CIRCLE|TRAVIS_TAG|TRAVIS|TRAVIS_REPO_|TRAVIS_BUILD_|TRAVIS_BRANCH|TRAVIS_PULL_REQUEST_|APPVEYOR_|CSC_|GH_|GITHUB_|BT_|AWS_|STRIP|BUILD_') \
   --env http_proxy \
   --env https_proxy \
   --env ELECTRON_CACHE="/root/.cache/electron" \
   --env ELECTRON_BUILDER_CACHE="/root/.cache/electron-builder" \
+  -v ${EXTENSION_PATH}:/extension \
   -v ${PWD}:/project \
   -v ${PWD##*/}-node-modules:/project/node_modules \
   -v ~/.cache/electron:/root/.cache/electron \
   -v ~/.cache/electron-builder:/root/.cache/electron-builder \
   electronuserland/builder:wine \
-  /bin/bash -c "yarn && npm run dist"
+  /bin/bash -c "npm ci && npm run dist"
